refactor(test/e2e): move login functionality to doLogin util
